refactor(interactions): move in its own package
