+++ examples/components/main.go
@@ -3,8 +3,11 @@ package main
 import (
 	"flag"
 	"fmt"
+
 	"github.com/nyttikord/gokord/channel"
 	"github.com/nyttikord/gokord/component"
+	"github.com/nyttikord/gokord/interactions"
+
 	"log"
 	"os"
 	"os/signal"
@@ -38,9 +41,9 @@ func init() {
 var (
 	componentsHandlers = map[string]func(s *gokord.Session, i *gokord.InteractionCreate){
 		"fd_no": func(s *gokord.Session, i *gokord.InteractionCreate) {
-			err := s.InteractionRespond(i.Interaction, &gokord.InteractionResponse{
+			err := s.InteractionRespond(i.Interaction, &interactions.InteractionResponse{
 				Type: gokord.InteractionResponseChannelMessageWithSource,
-				Data: &gokord.InteractionResponseData{
+				Data: &interactions.InteractionResponseData{
 					Content: "Huh. I see, maybe some of these resources might help you?",
 					Flags:   channel.MessageFlagsEphemeral,
 					Components: []component.Message{
@@ -80,9 +83,9 @@ var (
 			}
 		},
 		"fd_yes": func(s *gokord.Session, i *gokord.InteractionCreate) {
-			err := s.InteractionRespond(i.Interaction, &gokord.InteractionResponse{
+			err := s.InteractionRespond(i.Interaction, &interactions.InteractionResponse{
 				Type: gokord.InteractionResponseChannelMessageWithSource,
-				Data: &gokord.InteractionResponseData{
+				Data: &interactions.InteractionResponseData{
 					Content: "Great! If you wanna know more or just have questions, feel free to visit Discord Devs and Discord Gophers server. " +
 						"But now, when you know how buttons work, let's move onto select menus (execute `/selects single`)",
 					Flags: channel.MessageFlagsEphemeral,
@@ -115,22 +118,22 @@ var (
 			}
 		},
 		"select": func(s *gokord.Session, i *gokord.InteractionCreate) {
-			var response *gokord.InteractionResponse
+			var response *interactions.InteractionResponse
 
 			data := i.MessageComponentData()
 			switch data.Values[0] {
 			case "go":
-				response = &gokord.InteractionResponse{
+				response = &interactions.InteractionResponse{
 					Type: gokord.InteractionResponseChannelMessageWithSource,
-					Data: &gokord.InteractionResponseData{
+					Data: &interactions.InteractionResponseData{
 						Content: "This is the way.",
 						Flags:   channel.MessageFlagsEphemeral,
 					},
 				}
 			default:
-				response = &gokord.InteractionResponse{
+				response = &interactions.InteractionResponse{
 					Type: gokord.InteractionResponseChannelMessageWithSource,
-					Data: &gokord.InteractionResponseData{
+					Data: &interactions.InteractionResponseData{
 						Content: "It is not the way to go.",
 						Flags:   channel.MessageFlagsEphemeral,
 					},
@@ -155,9 +158,9 @@ var (
 
 			const stackoverflowFormat = `https://stackoverflow.com/questions/tagged/%s`
 
-			err := s.InteractionRespond(i.Interaction, &gokord.InteractionResponse{
+			err := s.InteractionRespond(i.Interaction, &interactions.InteractionResponse{
 				Type: gokord.InteractionResponseChannelMessageWithSource,
-				Data: &gokord.InteractionResponseData{
+				Data: &interactions.InteractionResponseData{
 					Content: "Here is your stackoverflow URL: " + fmt.Sprintf(stackoverflowFormat, strings.Join(data.Values, "+")),
 					Flags:   channel.MessageFlagsEphemeral,
 				},
@@ -175,9 +178,9 @@ var (
 			}
 		},
 		"channel_select": func(s *gokord.Session, i *gokord.InteractionCreate) {
-			err := s.InteractionRespond(i.Interaction, &gokord.InteractionResponse{
+			err := s.InteractionRespond(i.Interaction, &interactions.InteractionResponse{
 				Type: gokord.InteractionResponseChannelMessageWithSource,
-				Data: &gokord.InteractionResponseData{
+				Data: &interactions.InteractionResponseData{
 					Content: "This is it. You've reached your destination. Your choice was <#" + i.MessageComponentData().Values[0] + ">\n" +
 						"If you want to know more, check out the links below",
 					Components: []component.Message{
@@ -221,9 +224,9 @@ var (
 	}
 	commandsHandlers = map[string]func(s *gokord.Session, i *gokord.InteractionCreate){
 		"buttons": func(s *gokord.Session, i *gokord.InteractionCreate) {
-			err := s.InteractionRespond(i.Interaction, &gokord.InteractionResponse{
+			err := s.InteractionRespond(i.Interaction, &interactions.InteractionResponse{
 				Type: gokord.InteractionResponseChannelMessageWithSource,
-				Data: &gokord.InteractionResponseData{
+				Data: &interactions.InteractionResponseData{
 					Content: "Are you comfortable with buttons and other message components?",
 					Flags:   channel.MessageFlagsEphemeral,
 					// Buttons and other components are specified in Components field.
@@ -278,12 +281,12 @@ var (
 			}
 		},
 		"selects": func(s *gokord.Session, i *gokord.InteractionCreate) {
-			var response *gokord.InteractionResponse
+			var response *interactions.InteractionResponse
 			switch i.ApplicationCommandData().Options[0].Name {
 			case "single":
-				response = &gokord.InteractionResponse{
+				response = &interactions.InteractionResponse{
 					Type: gokord.InteractionResponseChannelMessageWithSource,
-					Data: &gokord.InteractionResponseData{
+					Data: &interactions.InteractionResponseData{
 						Content: "Now let's take a look on selects. This is single item select menu.",
 						Flags:   channel.MessageFlagsEphemeral,
 						Components: []component.Message{
@@ -331,9 +334,9 @@ var (
 				}
 			case "multi":
 				minValues := 1
-				response = &gokord.InteractionResponse{
+				response = &interactions.InteractionResponse{
 					Type: gokord.InteractionResponseChannelMessageWithSource,
-					Data: &gokord.InteractionResponseData{
+					Data: &interactions.InteractionResponseData{
 						Content: "Now let's see how the multi-item select menu works: " +
 							"try generating your own stackoverflow search link",
 						Flags: channel.MessageFlagsEphemeral,
@@ -398,9 +401,9 @@ var (
 					},
 				}
 			case "auto-populated":
-				response = &gokord.InteractionResponse{
+				response = &interactions.InteractionResponse{
 					Type: gokord.InteractionResponseChannelMessageWithSource,
-					Data: &gokord.InteractionResponseData{
+					Data: &interactions.InteractionResponseData{
 						Content: "The tastiest things are left for the end. Meet auto populated select menus.\n" +
 							"By setting `MenuType` on the select menu you can tell Discord to automatically populate the menu with entities of your choice: roles, members, channels. Try one below.",
 						Flags: channel.MessageFlagsEphemeral,
@@ -445,7 +448,7 @@ func main() {
 			}
 		}
 	})
-	_, err := s.ApplicationCommandCreate(*AppID, *GuildID, &gokord.ApplicationCommand{
+	_, err := s.ApplicationCommandCreate(*AppID, *GuildID, &interactions.Command{
 		Name:        "buttons",
 		Description: "Test the buttons if you got courage",
 	})
@@ -453,9 +456,9 @@ func main() {
 	if err != nil {
 		log.Fatalf("Cannot create slash command: %v", err)
 	}
-	_, err = s.ApplicationCommandCreate(*AppID, *GuildID, &gokord.ApplicationCommand{
+	_, err = s.ApplicationCommandCreate(*AppID, *GuildID, &interactions.Command{
 		Name: "selects",
-		Options: []*gokord.ApplicationCommandOption{
+		Options: []*interactions.CommandOption{
 			{
 				Type:        gokord.ApplicationCommandOptionSubCommand,
 				Name:        "multi",
